Move DbConnectionPoolGroup to Common
