+++ src/System.Data.SqlClient/src/System.Data.SqlClient.csproj
@@ -86,11 +86,13 @@
     <Compile Include="$(CommonPath)\System\Data\ProviderBase\DbConnectionFactory.cs">
       <Link>Common\System\Data\ProviderBase\DbConnectionFactory.cs</Link>
     </Compile>
+    <Compile Include="$(CommonPath)\System\Data\ProviderBase\DbConnectionPoolGroup.cs">
+      <Link>Common\System\Data\ProviderBase\DbConnectionPoolGroup.cs</Link>
+    </Compile>
     <Compile Include="System\Data\ProviderBase\DbConnectionClosed.cs" />
     <Compile Include="System\Data\ProviderBase\DbConnectionFactory.cs" />
     <Compile Include="System\Data\ProviderBase\DbConnectionInternal.cs" />
     <Compile Include="System\Data\ProviderBase\DbConnectionPool.cs" />
-    <Compile Include="System\Data\ProviderBase\DbConnectionPoolGroup.cs" />
     <Compile Include="System\Data\ProviderBase\DbConnectionPoolGroupProviderInfo.cs" />
     <Compile Include="System\Data\ProviderBase\DbConnectionPoolIdentity.cs" />
     <Compile Include="System\Data\ProviderBase\DbConnectionPoolOptions.cs" />
